% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParameters.R
\name{addParameters}
\alias{addParameters}
\title{Adding parameters}
\usage{
addParameters(x, pName = NA, iVal = "1", overwrite = FALSE)
}
\arguments{
\item{x}{model to which the parameter is added (required)}

\item{pName}{parameter name (required)}

\item{iVal}{parameter value; if the value is not specified, sysBio will set it to 1}

\item{overwrite}{a flag that allows changes to the existing parameter (default value FALSE)}
}
\value{
This function adds information about parameters into the model (given as a first argument of the function). 
    Parameters information is stored in the list format and contain the following elements:
    \itemize{
    \item{parameters$pName - parameter name}
    \item{parameters$initVal - parameter value}
    }
}
\description{
This function allows users to specify values of parameters used in the reaction (if parameters are used). The current 
sysBio version assumes that parameters, if defined, will remain constant during simulation. Thus, only numerical values 
can be assigned to parameter values.
}
\examples{
exmp <- newModel("This is an example of a new model")
addMAreaction(exmp, react="A = null", "rf", "rb")
addMAreaction(exmp, react="A + B -> 2*AB", "k", name="Forward AB")
addMAreaction(exmp, react="AB -> null", "rAB")

addMAreactRate(exmp, "rf", "fixed", "1")
addMAreactRate(exmp, "rb", "fixed", "0.75")
addMAreactRate(exmp, "k", "fixed", "0.5")
addMAreactRate(exmp, "rAB", "assigned", "p1*A")

addParameters(exmp, "p1", 0.75)
 
# Show info about model reactions and parameters
exmp$reaction
exmp$parameters

}

