% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRule.R
\name{addRule}
\alias{addRule}
\title{Adding rules}
\usage{
addRule(x, rName = NA, type = NA, rule = NA, overwrite = FALSE)
}
\arguments{
\item{x}{model to which the rule is added (required)}

\item{rName}{- rule name (required)}

\item{type}{rule type; currently only type ODEs is supported (rules in the form of ODEs) (required)}

\item{rule}{rule definition (required)}

\item{overwrite}{a flag that allows changes to the existing rule (default value FALSE)}
}
\value{
This function adds information about rules into the model (given as a first argument of the function). 
    Rules information is stored in the list format and contain the following elements:  
    \itemize{
    \item{rule$rName - rule name}
    \item{rule$rType - rule type (ODEs)}
    \item{rule$rRule - rule definition}
    }
}
\description{
This function allows user to define rules. Rules can be used to specify values for model species 
based on the values of other components (mainly species) in the model. Current version allows only 
for the ODEs type of rules.
}
\examples{
exmp <- newModel("This is an example of a new model")
addMAreaction(exmp, react="A = null", "rf", "rb")
addMAreaction(exmp, react="A + B -> 2*AB", "k", name="Forward AB")
addMAreaction(exmp, react="AB -> null", "rAB")

addMAreactRate(exmp, "rf", "fixed", "1")
addMAreactRate(exmp, "rb", "fixed", "0.75")
addMAreactRate(exmp, "k", "fixed", "0.5")
addMAreactRate(exmp, "rAB", "assigned", "p1*A")

addParameters(exmp, "p1", 0.75)
 
addSpecies(exmp, "A", 10)
addSpecies(exmp, "B", 10)
addSpecies(exmp, "AB", 0)

addRule(exmp, "rule B", "ODEs", "B=-0.1*AB")
  
# Show info about model reactions and rules
exmp$reaction
exmp$rule

}

