% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{Creating a model}
\usage{
makeModel(x)
}
\arguments{
\item{x}{model name (required)}
}
\value{
This function adds updates model (given as an argument 
    of the function) with information about mathematical model 
    of the system, ODEs, stochastic matrix, and propensity 
    function. This function also updates model's isChecked 
    flag, as model is created based on the most recent data. 
    It updates the following elements:
    \itemize{
    \item{model - mathematical model of the system}
    \item{odes$equations - a set of ODEs that correspond to the mathematical model of the system (in format for ODEs solver)}
    \item{odes$rates - reaction rates assignments (in format for ODEs solver)}
    \item{odes$parameters - parameters assignments (in format for ODEs solver)}
    \item{odes$rules - rules assignments (in format for ODEs solver)}
    \item{stochMatrix - stochastic matrix (rows correspond to species, columns to reactions)}
    \item{stochModel - propensity function (one equation for each reaction)}
    \item{isChecked - flag set to 1}
    }
}
\description{
This function creates mathematical model of the system, a set 
of differential equations that correspond to that mathematical 
model, stochastic matrix, and propensity function. It also 
transforms all defined rates, parameters, and rules into a 
format required by the ODEs solver function.
}
\examples{
exmp <- newModel("This is an example of a new model")
addMAreaction(exmp, react="A = null", "rf", "rb")
addMAreaction(exmp, react="A + B -> 2*AB", "k", name="Forward AB")
addMAreaction(exmp, react="AB -> null", "rAB")

addMAreactRate(exmp, "rf", "fixed", "1")
addMAreactRate(exmp, "rb", "fixed", "0.75")
addMAreactRate(exmp, "k", "fixed", "0.5")
addMAreactRate(exmp, "rAB", "assigned", "p1*A")

addParameters(exmp, "p1", 0.75)
 
addSpecies(exmp, "A", 10)
addSpecies(exmp, "B", 10)
addSpecies(exmp, "AB", 0)

addRule(exmp, "rule B", "ODEs", "B=-0.1*AB")

# Show info about the model
exmp

makeModel(exmp)
  
# Show info about the updated model
exmp

}

