% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeODEParameters.R
\name{makeODEParameters}
\alias{makeODEParameters}
\title{Creating representation for reaction parameters}
\usage{
makeODEParameters(x)
}
\arguments{
\item{x}{model name (required)}
}
\value{
This function returns model parameters in the format appropriate for the ODEs solver
}
\description{
This function creates assignment (in the format required for ODE solver) for each each of the specified 
paramterers based on their assigned/specified value. This function is called from the "makeModel.R" function.
}

