% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeODERules.R
\name{makeODERules}
\alias{makeODERules}
\title{Creating representation for the rules}
\usage{
makeODERules(x, tmpDF)
}
\arguments{
\item{x}{model name (required)}

\item{tmpDF}{a data frame that contains information about species and their mathematical representations
(temp data frame used within the "makeModel.R" function) (required)}
}
\value{
This function returns rules defined for the model in the format appropriate for the ODEs solver
}
\description{
This function  creates assignment (in the format required for ODE solver) for each of the specified rules
based on their defined values.  This function is called from the "makeModel.R" function.
}

