% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeODEs.R
\name{makeODEs}
\alias{makeODEs}
\title{Creating a residual function for the ODEs solver}
\usage{
makeODEs(x)
}
\arguments{
\item{x}{model name (required)}
}
\value{
A residual function that will be used within the ODEs solver.
}
\description{
This function  creates a residual function that will be used within the ODEs solver. 
This function is called from the "simulateModel.R" function
}

