% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeProduct.R
\name{makeProduct}
\alias{makeProduct}
\title{Creating a product for reaction species}
\usage{
makeProduct(dfTmp, whichOne)
}
\arguments{
\item{dfTmp}{a data frame that contains}

\item{whichOne}{a flag that determines whether to use the reactant/species (e.g., A) (flag set to 1) or 
a derivative of the reactant (e.g., dA) in the equation (flag != 1)}
}
\value{
This function returns a product form for a reactant/species from the reactions
}
\description{
This function creates the product form for all of the species. It is called from the "parseComponents.R" function.
}

