% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStMat.R
\name{makeStMat}
\alias{makeStMat}
\title{Creating a stochastic matrix}
\usage{
makeStMat(x, tmpDF)
}
\arguments{
\item{x}{model name (required)}

\item{tmpDF}{a data frame that contains information about reactions, species, stoichometric coefficients, etc. 
(temp data frame used within the "makeModel.R" function) (required)}
}
\value{
This function returns a stochastic matrix that corresponds to a defined model.
}
\description{
This function creates a stochastic matrix of the model (required for the stochastic simulation)
This function is called from the "makeModel.R" function.
}

