% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStModel.R
\name{makeStModel}
\alias{makeStModel}
\title{Creating a propensity function}
\usage{
makeStModel(x, tmpDF)
}
\arguments{
\item{x}{model name (required)}

\item{tmpDF}{a data frame that contains information about reactions, species, stoichometric coefficients, etc. 
(temp data frame used within the "makeModel.R" function) (required)}
}
\value{
This function returns a vector that contains propensity function corresponding to a defined model.
}
\description{
This function creates a propensity function of the model (required for the stochastic simulation)
This function is called from the "makeModel.R" function.
}

