% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newModel.R
\name{newModel}
\alias{newModel}
\title{Creating a new model}
\usage{
newModel(x)
}
\arguments{
\item{x}{name of the model (required)}
}
\value{
This function adds information about rules into the model (given as a first argument of the function). 
    Rules information is stored in the list format and contain the following elements:  
    \itemize{
    \item{modelName - name of the model}
    \item{reaction - list of reactions (empty)}
    \item{species - list of species (empty)}
    \item{rates - list of reaction rates (empty)}
    \item{parameters - list of reaction parameters (empty)}
    \item{events - list of events (empty)}
    \item{rules - list of rules (empty)}
    \item{model - mathematical model of the system (empty)}
    \item{odes - mathematical model of the system in ODEs format (empty)}
    \item{stochMatrix - stochastic matrix (empty)}
    \item{stochModel - stochastic model (propensity function) (empty)}
    \item{isChecked - a flag that marks that changes have been done to the model, e.g., new reaction or species added (set to value 0)}
    }
}
\description{
Creating a new model object. This object will contain all 
components of the model and information required to run 
simulation.
}
\examples{
exmp <- newModel("This is an example of a new model")

# Show info about model
exmp

}

