% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseComponents.R
\name{parseComponents}
\alias{parseComponents}
\title{Parsing reaction components}
\usage{
parseComponents(model, components, rate)
}
\arguments{
\item{model}{model name (required)}

\item{components}{a vector that contain left and right side of the reaction (required)}

\item{rate}{reaction rate (required)}
}
\value{
This function returns a data frame that contains parsed components of a reaction.
}
\description{
This function parses components of a reaction. This function is called from the "parseReaction.R" function.
}

