% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseReaction.R
\name{parseReaction}
\alias{parseReaction}
\title{Parsing reactions}
\usage{
parseReaction(model, reaction, rate, rNumber)
}
\arguments{
\item{model}{model name (required)}

\item{reaction}{reaction (forward reaction only) (required)}

\item{rate}{reaction rate (required)}

\item{rNumber}{number of reaction in the model; if reaction is reversible, this number will be defined in the following format:
reactionNumber_1 and reactionNumber_2 (required)}
}
\value{
This function returns a data frame that contains parsed reaction.
}
\description{
This function parses eaction and transforms the canonical form of a mass action reaction 
into a format that can be used to describe them as ODEs. This function is called from the
"parseReaction.R" function.
}

