% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResults}
\alias{plotResults}
\title{Results visualization}
\usage{
plotResults(x = NULL, smooth = FALSE, title = NULL)
}
\arguments{
\item{x}{a data frame that contains results of the simulation (required)}

\item{smooth}{a flag that specify type of geom(s) to draw - smooth (flag set to TRUE) or line (flag set to FALSE).}

\item{title}{figure title}
}
\value{
Figure
}
\description{
This function provides a simple visualization of simulation results, where time points are plotted on the x-axis 
and the corresponding species amounts for each species are plotted on the y-axis.
}
\examples{
exmp <- newModel("This is an example of a new model")
addMAreaction(exmp, react="A = null", "rf", "rb")
addMAreaction(exmp, react="A + B -> 2*AB", "k", name="Forward AB")
addMAreaction(exmp, react="AB -> null", "rAB")

addMAreactRate(exmp, "rf", "fixed", "1")
addMAreactRate(exmp, "rb", "fixed", "0.75")
addMAreactRate(exmp, "k", "fixed", "0.5")
addMAreactRate(exmp, "rAB", "assigned", "p1*A")

addParameters(exmp, "p1", 0.75)
 
addSpecies(exmp, "A", 10)
addSpecies(exmp, "B", 10)
addSpecies(exmp, "AB", 0)

addRule(exmp, "rule B", "ODEs", "B=-0.1*AB")

makeModel(exmp)
  
simResults <-simulateModel(exmp)
plotResults(simResults, title="Simulation results")

}

