% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printInfo.R
\name{printInfo}
\alias{printInfo}
\title{Print information about the model}
\usage{
printInfo(x, allDetails = FALSE)
}
\arguments{
\item{x}{a model name}

\item{allDetails}{a flag that specify whether the user wants short or long description of the model (flag set to FALSE).}
}
\value{
Textual summary
}
\description{
This function prints out info about mode: model name, number of reactions and their types, number of species, and number of ODEs required to
represent the model. It also allows users to print detailed info about each of the reactions, species, reaction rates, parameters, rules, 
and model equations.
}
\examples{
exmp <- newModel("This is an example of a new model")
addMAreaction(exmp, react="A = null", "rf", "rb")
addMAreaction(exmp, react="A + B -> 2*AB", "k", name="Forward AB")
addMAreaction(exmp, react="AB -> null", "rAB")

addMAreactRate(exmp, "rf", "fixed", "1")
addMAreactRate(exmp, "rb", "fixed", "0.75")
addMAreactRate(exmp, "k", "fixed", "0.5")
addMAreactRate(exmp, "rAB", "assigned", "p1*A")

addParameters(exmp, "p1", 0.75)
 
addSpecies(exmp, "A", 10)
addSpecies(exmp, "B", 10)
addSpecies(exmp, "AB", 0)

addRule(exmp, "rule B", "ODEs", "B=-0.1*AB")

makeModel(exmp)
  
printInfo(exmp)
printInfo(exmp, allDetails=TRUE)

}

