% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateModel.R
\name{simulateModel}
\alias{simulateModel}
\title{Model simulation (ordinary differential equation solver)}
\usage{
simulateModel(x, times = seq(0, 10, by = 0.1))
}
\arguments{
\item{x}{model name (required)}

\item{times}{time points in which model will be simulated}
}
\value{
Data frame where each row corresponds to a time point in which model was simulated. First column corresponds to 
    time, while the rest of columns correspond to model species, i.e., values assigned to those species at a given time point.
}
\description{
This function allows users to simulate the model (i.e., solve ODEs) using a differential equation solver 
based on a combination of backward differentiation formula and a direct linear system solution method 
(using the "daspk" function from the deSolve package).
}
\examples{
exmp <- newModel("This is an example of a new model")
addMAreaction(exmp, react="A = null", "rf", "rb")
addMAreaction(exmp, react="A + B -> 2*AB", "k", name="Forward AB")
addMAreaction(exmp, react="AB -> null", "rAB")

addMAreactRate(exmp, "rf", "fixed", "1")
addMAreactRate(exmp, "rb", "fixed", "0.75")
addMAreactRate(exmp, "k", "fixed", "0.5")
addMAreactRate(exmp, "rAB", "assigned", "p1*A")

addParameters(exmp, "p1", 0.75)
 
addSpecies(exmp, "A", 10)
addSpecies(exmp, "B", 10)
addSpecies(exmp, "AB", 0)

addRule(exmp, "rule B", "ODEs", "B=-0.1*AB")

makeModel(exmp)
  
simResults <-simulateModel(exmp)
plotResults(simResults, title="Simulation results")

}

