% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveStoch.R
\name{solveStoch}
\alias{solveStoch}
\title{Model simulation (stochastic simulation)}
\usage{
solveStoch(x, times = 10, method = "D", simName = "", tau = 0.3,
  f = 10, epsilon = 0.03, nc = 10, hor = NaN, dtf = 10, nd = 100,
  ignoreNegativeState = TRUE, consoleInterval = 0, censusInterval = 0,
  verbose = FALSE, maxWallTime = Inf)
}
\arguments{
\item{x}{model name (required)}

\item{times}{time points in which model will be simulated}

\item{method}{text string indicating the SSA method to use (see ?GillespieSSA::ssa for more details)}

\item{simName}{see ?GillespieSSA::ssa for more details}

\item{tau}{see ?GillespieSSA::ssa for more details}

\item{f}{see ?GillespieSSA::ssa for more details}

\item{epsilon}{see ?GillespieSSA::ssa for more details}

\item{nc}{see ?GillespieSSA::ssa for more details}

\item{hor}{see ?GillespieSSA::ssa for more details}

\item{dtf}{see ?GillespieSSA::ssa for more details}

\item{nd}{see ?GillespieSSA::ssa for more details}

\item{ignoreNegativeState}{see ?GillespieSSA::ssa for more details}

\item{consoleInterval}{see ?GillespieSSA::ssa for more details}

\item{censusInterval}{see ?GillespieSSA::ssa for more details}

\item{verbose}{see ?GillespieSSA::ssa for more details}

\item{maxWallTime}{see ?GillespieSSA::ssa for more details}
}
\value{
Data frame where each row corresponds to a time point in which model was simulated. First column corresponds to 
    time, while the rest of columns correspond to model species, i.e., values assigned to those species at a given time point.
}
\description{
This function allows users to simulate the model (i.e., solve ODEs) using Gillespie stochastic simulation algorithm (using the "ssa" function from the GillespieSSA package).
}
\examples{
exmp <- newModel("This is an example of a new model")
addMAreaction(exmp, react="A = null", "rf", "rb")
addMAreaction(exmp, react="A + B -> 2*AB", "k", name="Forward AB")
addMAreaction(exmp, react="AB -> null", "rAB")

addMAreactRate(exmp, "rf", "fixed", "1")
addMAreactRate(exmp, "rb", "fixed", "0.75")
addMAreactRate(exmp, "k", "fixed", "0.5")
addMAreactRate(exmp, "rAB", "assigned", "p1*A")

addParameters(exmp, "p1", 0.75)
 
addSpecies(exmp, "A", 10)
addSpecies(exmp, "B", 10)
addSpecies(exmp, "AB", 0)

addRule(exmp, "rule B", "ODEs", "B=-0.1*AB")

makeModel(exmp)
  
simResults.stoch <- solveStoch(exmp)
plotResults(simResults.stoch, title="Simulation results (stochastic)")

}

