% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesDetails.R
\name{speciesDetails}
\alias{speciesDetails}
\title{Get info abour species included in the reaction}
\usage{
speciesDetails(x)
}
\arguments{
\item{x}{a model name}
}
\value{
A vector that contains species name and the corresponding stoichometric coefficient
}
\description{
This function returns info about species included in the reaction (species name and the corresponding stoichometric coefficient). 
It is called within the "parseComponents.R" function
}

