% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateModel.R
\name{validateModel}
\alias{validateModel}
\title{Print information about the model}
\usage{
validateModel(x)
}
\arguments{
\item{x}{a model name}
}
\value{
Numeric value - 1 if model has been validated, 0 if validation failed
}
\description{
This function validates model, i.e., checks if all reactions, parameters, rates, rules, etc., are 
incoorporated into a model, as well as whether all variables have assigned values (either constant or in a form of a function).
}
\examples{
exmp <- newModel("This is an example of a new model")
addMAreaction(exmp, react="A = null", "rf", "rb")
addMAreaction(exmp, react="A + B -> 2*AB", "k", name="Forward AB")
addMAreaction(exmp, react="AB -> null", "rAB")

addMAreactRate(exmp, "rf", "fixed", "1")
addMAreactRate(exmp, "rb", "fixed", "0.75")
addMAreactRate(exmp, "k", "fixed", "0.5")
addMAreactRate(exmp, "rAB", "assigned", "p1*A")

addParameters(exmp, "p1", 0.75)
 
addSpecies(exmp, "A", 10)
addSpecies(exmp, "B", 10)
addSpecies(exmp, "AB", 0)

addRule(exmp, "rule B", "ODEs", "B=-0.1*AB")

makeModel(exmp)
  
validateModel(exmp)

}

